pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.38"}
alembic = "^1.14.1"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
pydantic-settings = "^2.8.0"
asyncpg = "^0.30.0"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.3.4"

